+++ crates/plugin_partial_bundling/src/lib.rs
@@ -1,6 +1,8 @@
-use std::collections::HashSet;
+use std::collections::{HashMap, HashSet};
+use std::sync::RwLock;
 use std::{collections::VecDeque, sync::Arc};
 
+use farmfe_core::config::custom::CUSTOM_CONFIG_PARTIAL_BUNDLING_GROUPS_ENFORCE_MAP;
 use farmfe_core::{
   config::Config,
   context::CompilationContext,
@@ -25,9 +27,24 @@ mod module_pot;
 mod utils;
 /// Partial Bundling implementation for Farm.
 /// See https://github.com/farm-fe/rfcs/pull/9
-pub struct FarmPluginPartialBundling {}
+pub struct FarmPluginPartialBundling {
+  partial_bundling_groups_enforce_map: RwLock<HashMap<String, bool>>,
+}
 
 impl Plugin for FarmPluginPartialBundling {
+  fn config(&self, config: &mut Config) -> farmfe_core::error::Result<Option<()>> {
+    *self.partial_bundling_groups_enforce_map.write().unwrap() = config
+      .custom
+      .get(CUSTOM_CONFIG_PARTIAL_BUNDLING_GROUPS_ENFORCE_MAP)
+      .map(|s| {
+        farmfe_core::serde_json::from_str(s)
+          .expect("failed to parse partial bundling group enforce map")
+      })
+      .unwrap_or_default();
+
+    Ok(Some(()))
+  }
+
   fn name(&self) -> &str {
     "FarmPluginPartialBundling"
   }
@@ -74,12 +91,15 @@ impl Plugin for FarmPluginPartialBundling {
     let module_group_buckets =
       group_module_buckets_by_module_group(&module_buckets_map, &module_group_graph, &module_graph);
 
+    let enforce_map = self.partial_bundling_groups_enforce_map.read().unwrap();
+
     // 3. generate resource pots
     let resource_pots = generate_resource_pots(
       module_group_buckets,
       module_buckets_map,
       &module_graph,
-      &context.config.partial_bundling,
+      &context.config,
+      &enforce_map,
     );
 
     Ok(Some(resource_pots))
@@ -88,7 +108,9 @@ impl Plugin for FarmPluginPartialBundling {
 
 impl FarmPluginPartialBundling {
   pub fn new(_: &Config) -> Self {
-    Self {}
+    Self {
+      partial_bundling_groups_enforce_map: RwLock::new(HashMap::new()),
+    }
   }
 }
 
@@ -226,7 +248,9 @@ mod tests {
 
   #[test]
   fn analyze_module_graph() {
-    let plugin = FarmPluginPartialBundling {};
+    let plugin = FarmPluginPartialBundling {
+      partial_bundling_groups_enforce_map: Default::default(),
+    };
     let mut context = CompilationContext::new(Default::default(), vec![]).unwrap();
     let graph = construct_test_module_graph();