+++ crates/toolkit/src/common.rs
@@ -33,7 +33,7 @@ pub struct Source {
 /// create a swc source map from a source
 pub fn create_swc_source_map(source: Source) -> (Arc<SourceMap>, Arc<SourceFile>) {
   let cm = Arc::new(SourceMap::default());
-  let sf = cm.new_source_file_from(FileName::Real(source.path), source.content);
+  let sf = cm.new_source_file_from(Arc::new(FileName::Real(source.path)), source.content);
 
   (cm, sf)
 }
@@ -181,9 +181,7 @@ pub fn load_source_original_source_map(
   let mut map = None;
   // try load source map when load module content.
   if content.contains(source_map_comment_prefix) {
-    let base64_prefix = format!(
-      "{source_map_comment_prefix}=data:application/json;base64,"
-    );
+    let base64_prefix = format!("{source_map_comment_prefix}=data:application/json;base64,");
     // detect that the source map is inline or not
     let source_map = if content.contains(&base64_prefix) {
       // inline source map