+++ examples/vite-adapter-solid/src/App.tsx
@@ -15,7 +15,7 @@ const App: Component = () => {
       <header class={styles.header}>
         <img src={logo} class={styles.logo} alt="logo" />
         <p>
-          Edit<code>src/App.tsx</code> and save to reload.
+          Edit<code>src/App.tsx</code>1232131231231221312321 and save to reload.
         </p>
         <a
           class={styles.link}